refactored code and comments
