+++ v2/elasticsearch-common/src/main/java/com/google/cloud/teleport/v2/elasticsearch/utils/Dataset.java
@@ -30,4 +30,11 @@ public enum Dataset {
     public String getKey() {
         return name().toLowerCase();
     }
+
+    /**
+     * This method is used to get Dataset literal with the "gcp." prefix.
+     * **/
+    public String getKeyWithPrefix() {
+        return "gcp." + name().toLowerCase();
+    }
 }
