+++ v2/elasticsearch-common/src/main/java/com/google/cloud/teleport/v2/elasticsearch/utils/ElasticsearchIO.java
@@ -1050,6 +1050,9 @@ public class ElasticsearchIO {
      */
     public interface FieldValueExtractFn extends SerializableFunction<JsonNode, String> {}
 
+    /**
+     * BooleanFieldValueExtractFn class.
+     */
     public interface BooleanFieldValueExtractFn extends SerializableFunction<JsonNode, Boolean> {}
 
     abstract @Nullable ConnectionConfiguration getConnectionConfiguration();
