+++ v2/elasticsearch-common/src/main/java/com/google/cloud/teleport/v2/elasticsearch/utils/ElasticsearchIndex.java
@@ -28,6 +28,6 @@ public class ElasticsearchIndex {
     }
 
     public String getIndex(){
-        return "logs-gcp." + this.dataset.toString().toLowerCase() + "-" + this.namespace;
+        return "logs-" + this.dataset.toString().toLowerCase() + "-" + this.namespace;
     }
 }
